[addons] add initial new add-on header 'AddonBase.h'
This is the coming main add-on header to handle the base parts who are
instance independent.
